Move TableInitializer to wasm_common
